requirements.txt
# Core
python-dotenv>=1.0.0

# Scraping
praw>=7.7.0
feedparser>=6.0.0
scrapetube>=2.5.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# AI
openai>=1.12.0

# Technical Analysis
pandas>=2.2.0
ta>=0.11.0
numpy>=1.26.0

# Delivery
python-telegram-bot>=21.0

# Utilities
requests>=2.31.0
pytz>=2024.1
